requirements.txt
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
